refactor(rewrite): switch to use dush-better-use
it includes minibase-is-registered itself.

BREAKING CHANGE: Refactor. Now it returns a function that accepts optional options object, when
called it retusn a function that should be passed to .use method, it is just a convention plugins to
be written this way - function that accepts options (optional) and returns a plugin function. Like:
createPlugin([name,] fn) -> plugin; plugin([options]) -> fn; app.use(fn)